move tournament challenges logic into a concern
